==> nb/2.compute_rp.ipynb <==
import numpy as np
import pandas as pd
import statsmodels.api as sm
import tqdm.notebook

# Computing phenotypic correlations

This is loosely based on https://gist.github.com/ce-carey/6480d6544f132829d9579b2a1f1455b4.

We will compute the correlations between residualized phenotypes after adjusting for the GWAS covariates.

# 1. Build outcomes (phenotypes) table

relevant_icd_10_df = pd.read_csv('../data/phenotypes_meta/chosen_icd_codes.csv', header=None, names=['code'])

relevant_icd_10_df.head(2)

# Copy the table to the local directory
# icd_10_df = pd.read_csv('/data1/deep_storage/ukbiobank/ukb_datapulls/ukb23674_icd10_long.txt.gz', 
#                         sep='\t', compression='gzip')

# icd_10_df.to_csv('../data/phenotypes/ukb_raw_icd10_long.tsv.gz', sep='\t', index=False, compression='gzip')

icd_10_df = pd.read_csv('../data/phenotypes/ukb_raw_icd10_long.tsv.gz', sep='\t', compression='gzip')

icd_10_df.head(0)

all_eid_df = icd_10_df[['eid']].drop_duplicates()

all_eid_df.to_csv('../data/phenotypes/all_eids.tsv', sep='\t', index=False)

all_eid_df.head(0)

relevant_code_occurrences_df = (
    icd_10_df
    .assign(
        code=lambda df: df['icd10_code'].apply(lambda x: x[:3]),
        had_code=1
    )
    .merge(relevant_icd_10_df, on='code')
    .groupby(['eid', 'code'])
    .agg({'had_code': 'max'})
    .reset_index()
    .query('had_code == 1')
    .filter(items=['eid', 'code'])
)

relevant_code_occurrences_df.to_csv('../data/phenotypes/relevant_occurrences_long.tsv', sep='\t', index=False)

relevant_code_occurrences_df.head(0)

wide_phenotypes_df = (
    all_eid_df
    .merge(relevant_code_occurrences_df.assign(had=1), on='eid', how='left')
    .pivot_table(index='eid', columns='code', values='had', aggfunc='max', fill_value=0)
    .reset_index()
    .rename(columns={'eid': 'IID'})
)

wide_phenotypes_df.to_csv('../data/phenotypes/relevant_occurrences_wide.tsv', sep='\t', index=False)

wide_phenotypes_df.head(0)

# 2. Build covariates table

# # Copy the covariates file to the local directory and add a header
# covar_df = pd.read_csv('/data1/deep_storage/ukbiobank/pt2281/covar_chip/covar_chr1_chip.txt', 
#                        sep='\s+', header=None, usecols=list(range(14)), 
#                        names=['FID', 'IID', 'sex', 'age', 'PC1', 'PC2', 'PC3', 'PC4', 'PC5', 
#                               'PC6', 'PC7', 'PC8', 'PC9', 'PC10'])

# covar_df.to_csv('../data/phenotypes/raw_covariates.covar', sep='\t', index=False)

covar_df = pd.read_csv('../data/phenotypes/raw_covariates.covar', sep='\t')

covar_df.head(0)

gwas_covar_df = (
    covar_df
    .assign(
        age_squared=lambda df: df['age']**2,
        age_sex=lambda df: df['sex'] * df['age'],
        age_squared_sex=lambda df: df['sex'] * df['age_squared']
    )
)

gwas_covar_df.to_csv('../data/phenotypes/gwas_covariates.covar', sep='\t', index=False)

gwas_covar_df.head(0)

# 3. Compute residualized phenotypes

full_df = wide_phenotypes_df.merge(gwas_covar_df, on='IID').drop(columns=['FID'])

full_df.head(0)

phenotype_codes = relevant_icd_10_df['code'].values.tolist()

X = full_df[['PC1', 'PC2', 'PC3', 'PC4', 'PC5', 'PC6', 'PC7', 'PC8', 'PC9', 'PC10', 
             'age', 'sex', 'age_squared', 'age_sex', 'age_squared_sex']]

resid_df = full_df[['IID'] + phenotype_codes].copy()

for phenotype_code in tqdm.notebook.tqdm(phenotype_codes):
    reg = sm.OLS(full_df[phenotype_code].values, X)
    res = reg.fit()
    resid_df[phenotype_code] = res.resid

resid_df.to_csv('../data/phenotypes/residualized_phenotypes.tsv.gz', sep='\t', index=False, 
                compression='gzip')
    
resid_df.head(0)

# 4. Compute correlation between residualized phenotypes

correlation_df = (
    resid_df
    .drop(columns=['IID'])
    .corr()
    .stack()
    .reset_index()
    .rename(columns={'level_0': 'p1', 'level_1': 'p2', 0: 'rp'})
    .query('p1 < p2')
)

def normalize(values):
    return (values - values.mean()) / values.std()

covariance_df = (
    resid_df
    .drop(columns=['IID'])
    .apply(normalize, axis=0)
    .cov()
    .stack()
    .reset_index()
    .rename(columns={'level_0': 'p1', 'level_1': 'p2', 0: 'cp'})
    .query('p1 < p2')
)

phenotypic_correlation_df = correlation_df.merge(covariance_df, on=['p1', 'p2'])

phenotypic_correlation_df.to_csv('../data/phenotypes_meta/phenotypic_correlation.tsv', sep='\t', index=False)

phenotypic_correlation_df.head(2)
