==> nb/3.compute_coheritability.ipynb <==
import pathlib

import pandas as pd

codes_df = pd.read_csv('../data/phenotypes_meta/chosen_icd_codes.csv', header=None, names=['code'])

codes_df.head(2)

# Heritability information

h2_df = pd.read_csv('../data/neale_heritability_estimates/ukb31063_h2_all.02Oct2019.tsv.gz', 
                    usecols=['phenotype', 'sex', 'n_cases', 'n_controls', 'h2_liability', 
                             'h2_liability_se', 'h2_z', 'h2_p'], sep='\t', compression='gzip')

h2_df.head(2)

relevant_h2_df = (
    h2_df
    .query('sex == "both_sexes"')
    .merge(codes_df, left_on='phenotype', right_on='code')
    .filter(items=['code', 'n_cases', 'n_controls', 'h2_liability', 'h2_liability_se', 'h2_z', 'h2_p'])
)

relevant_h2_df.to_csv('../data/phenotypes_meta/heritability.tsv', sep='\t', index=False)

relevant_h2_df.head(2)

# Genetic correlation information

rg_df = pd.read_csv('../data/phenotypes_meta/genetic_correlation.tsv', sep='\t')

rg_df.head(2)

# Phenotypic correlation information

rp_df = pd.read_csv('../data/phenotypes_meta/phenotypic_correlation.tsv', sep='\t')

rp_df.head(2)

# Combine information

coheritability_df = (
    rg_df
    .merge(rp_df, on=['p1', 'p2'])
    .merge(relevant_h2_df, left_on='p1', right_on='code')
    .merge(relevant_h2_df, left_on='p2', right_on='code', suffixes=['_1', '_2'])
    .assign(coheritability=lambda df: (df['h2_liability_1'] * df['h2_liability_2'])**0.5 * df['rg'] / df['rp'])
    .filter(items=['code_1', 'code_2', 'h2_liability_1', 'h2_liability_2', 'rg', 'rp', 'coheritability'])
)

coheritability_df.to_csv('../data/phenotypes_meta/coheritability.tsv', sep='\t', index=False)

coheritability_df.head(2)
