==> nb/4.build_matrices.ipynb <==
import numpy as np
import pandas as pd
import torch

def make_square(df):
    df = df.copy()
    index = df.index.union(df.columns)
    df = df.reindex(index=index, columns=index)
    return df

def make_symmetric(df):
    df = df.copy().fillna(0)
    diag = np.diag(np.diag(df))
    df = df + df.T - diag
    return df

def fill_diag(df, value):
    df = df.copy()
    np.fill_diagonal(df.values, value)
    return df

coheritability_df = pd.read_csv('../data/phenotypes_meta/coheritability.tsv', sep='\t')

coheritability_df.head(2)

h2_df = pd.read_csv('../data/phenotypes_meta/heritability.tsv', sep='\t', usecols=['code', 'h2_liability'])

h2_df.head(2)

# Heritabilities are the main diagonal of the genetic covariance matrix
diagonal_df = (
    h2_df
    .merge(h2_df, on='code')
    .assign(code_2=lambda df: df['code'])
    .rename(columns={'code': 'code_1', 'h2_liability_x': 'rg'})
    .filter(items=['code_1', 'code_2', 'rg'])
)

diagonal_df.head(2)

genetic_covariance_matrix_df = (
    pd.concat([coheritability_df[['code_1', 'code_2', 'rg']], diagonal_df], ignore_index=True)
    .pivot(index='code_1', columns='code_2', values='rg')
    .pipe(make_square)
    .pipe(make_symmetric)
)

genetic_covariance_matrix_df.to_csv('../data/qtphenproxy_data/genetic_covariance_matrix.tsv', sep='\t')

genetic_covariance_matrix_df.iloc[:5, :5]

phenotypic_covariance_matrix_df = (
    coheritability_df
    .pivot(index='code_1', columns='code_2', values='rp')
    .pipe(make_square)
    .pipe(make_symmetric)
    .pipe(fill_diag, 1)
)

phenotypic_covariance_matrix_df.to_csv('../data/qtphenproxy_data/phenotypic_covariance_matrix.tsv', sep='\t')

phenotypic_covariance_matrix_df.iloc[:5, :5]
