==> nb/5.train_models.ipynb <==
import numpy as np
import pandas as pd
import torch
import tqdm

from plotnine import *

import sys

sys.path.append('../qtphenproxy_h2/')

import model

target_to_icd10 = {'stroke': 'I63', 'MI': 'I21'}

heritability_df = pd.read_csv('../data/phenotypes_meta/heritability.tsv', sep='\t')

heritability_df.head(2)

genetic_covariance_matrix_df = pd.read_csv('../data/qtphenproxy_data/genetic_covariance_matrix.tsv',
                                           sep='\t', index_col=0)

genetic_covariance_matrix_df.iloc[:3, :3]

phenotypic_covariance_matrix_df = pd.read_csv('../data/qtphenproxy_data/phenotypic_covariance_matrix.tsv',
                                              sep='\t', index_col=0)

phenotypic_covariance_matrix_df.iloc[:3, :3]

phenotypes_data_df = pd.read_csv('../data/phenotypes/relevant_occurrences_wide.tsv', 
                                 sep='\t', index_col=0)

phenotypes_data_df.head(0)

# Train model

def train(code, learning_rate=0.001, n_iter=5000, name=None):
    """Simple helper function for running this multiple times"""
    fitter = model.CombinationFitter.from_tables(code, genetic_covariance_matrix_df, 
                                                 phenotypic_covariance_matrix_df, phenotypes_data_df)
    fitter.fit(n_iter=n_iter, seed=0, learning_rate=learning_rate, verbose=False)

    if name is None:
        name = f'saved_qtphenproxy_{code}'
    fitter.save_fit(f'../data/{name}', person_ids=phenotypes_data_df.index.tolist())
    return fitter

mod1 = train('I21')

mod2 = train('I63')
