==> nb/6.gwas.ipynb <==
library(tidyverse)
library(broom)

# Load data files

covar_df <- read_tsv('../data/phenotypes/gwas_covariates.covar', col_types = cols())

covar_df %>% nrow

covar_df %>% head(0)

phenotypes_df <- read_tsv('../data/saved_qtphenproxy_I63/predictions.pheno', col_types = cols())

phenotypes_df %>% nrow

phenotypes_df %>% head(0)

genotypes_df <- read_tsv('../data/genotypes/marker_genotypes.raw', col_types = cols())

genotypes_df %>% nrow

genotypes_df %>% head(0)

# Run regressions

gwas_data_df <- genotypes_df %>% 
    inner_join(phenotypes_df, by = c('FID', 'IID')) %>%
    inner_join(covar_df, by = c('FID', 'IID'))

gwas_data_df %>% nrow

gwas_data_df %>% head(0)

snp_ids <- genotypes_df %>% select(-FID, -IID, -PAT, -MAT, -SEX, -PHENOTYPE) %>% colnames

snp_ids %>% length

gwas_results_df <- data.frame()

for (outcome in c('target', 'qtphenproxy', 'qtphenproxy_h2')) {
    if (outcome == 'target') {
        family = binomial
    } else {
        family = gaussian
    }
    
    for (snp_id in snp_ids) {
        reg_form <- paste(outcome, '~', 
                          paste(snp_id, 'sex', 'age', 'PC1', 'PC2', 'PC3', 'PC4', 'PC5', 'PC6', 'PC7', 
                                        'PC8', 'PC9', 'PC10', 'age_squared', 'age_sex', 'age_squared_sex', sep='+'))
        reg_rows <- glm(reg_form, data = gwas_data_df, family = family) %>% 
            tidy %>%
            mutate(outcome = outcome, snp_id = snp_id)
        
        gwas_results_df <- bind_rows(gwas_results_df, reg_rows)
    }
}

gwas_results_df %>% write_tsv("../data/gwas_results/I63.tsv")

gwas_results_df %>% head(2)

gwas_results_df %>%
    filter(str_detect(term, 'rs')) %>%
    mutate(significant = p.value < 0.05) %>%
    group_by(outcome) %>%
    summarize(n_significant = sum(significant), .groups = 'drop')
